[SpecModel] Move SpecModelError to separate file, document properties
